+++ mobile/lib/features/voice_hub/data/models/chat_service_message.dart
@@ -1,4 +1,4 @@
-import '../../core/util/json_util.dart';
+import '../../../../core/util/json_util.dart';
 
 class ChatServiceUser {
   const ChatServiceUser({